Move test_replication*
